refactor controller (New class Home)
